Add first test for JSON building
