+++ firmware/components/midi_mapping/CMakeLists.txt
@@ -1,13 +1,16 @@
 if(NOT DEFINED TESTING)
     idf_component_register(SRCS src/midi_mapping.cpp
+                                src/midi_mapping_json_parser.cpp
                            INCLUDE_DIRS include
                            REQUIRES etl audio_param midi rapidjson
                            PRIV_REQUIRES compiler_warning_flags)
 else()
     add_executable(midi_mapping_test
                    src/midi_mapping.cpp
+                   src/midi_mapping_json_parser.cpp
                    test/test_midi_mapping.cpp
                    test/test_json_parser.cpp
+                   test/test_json_builder.cpp
                    )
 
     target_include_directories(midi_mapping_test PRIVATE include)
