+++ firmware/components/midi_mapping/include/midi_mapping.h
@@ -170,6 +170,7 @@
  * </table>
  *
  */
+#pragma once
 
 #include <array>
 #include <cstdint>
